+++ interface/src/lib/components/GithubUpdateDialog.svelte
@@ -11,35 +11,31 @@
 	let updating = true;
 
 	let progress = 0;
-	$: if ($telemetry.github_update.status == 'progress') {
-		progress = $telemetry.github_update.progress;
+	$: if ($telemetry.download_ota.status == 'progress') {
+		progress = $telemetry.download_ota.progress;
 	}
 
-	$: if ($telemetry.github_update.status == 'error') {
+	$: if ($telemetry.download_ota.status == 'error') {
 		updating = false;
 	}
 
 	let message = 'Preparing ...';
 	let timerId: number;
 
-	$: if ($telemetry.github_update.status == 'progress') {
+	$: if ($telemetry.download_ota.status == 'progress') {
 		message = 'Downloading ...';
-	} else if ($telemetry.github_update.status == 'error') {
-		message = $telemetry.github_update.error;
-	} else if ($telemetry.github_update.status == 'finished') {
+	} else if ($telemetry.download_ota.status == 'error') {
+		message = $telemetry.download_ota.error;
+	} else if ($telemetry.download_ota.status == 'finished') {
 		// Catch interval triggering reactive statement over and over
 		if (message != 'Restarting ...') {
 			message = 'Restarting ...';
 			progress = 0;
 			// Reload page after 5 sec
-			timerId = setInterval(() => {
-				progress += 2;
-				if (progress >= 100) {
-					clearInterval(timerId);
-					closeAllModals();
-					location.reload();
-				}
-			}, 100);
+			timerId = setTimeout(() => {
+				closeAllModals();
+				location.reload();
+			}, 5000);
 		}
 	}
 
@@ -63,11 +59,15 @@
 		<div
 			class="bg-base-100 shadow-secondary/30 rounded-box pointer-events-auto flex max-h-full min-w-fit max-w-md flex-col justify-between p-4 shadow-lg"
 		>
-			<h2 class="text-base-content text-start text-2xl font-bold">Updating Firmware ...</h2>
+			<h2 class="text-base-content text-start text-2xl font-bold">Updating Firmware</h2>
 			<div class="divider my-2" />
 			<div class="overflow-y-auto">
 				<div class="bg-base-100 flex flex-col items-center justify-center p-6">
-					<progress class="progress progress-primary w-56" value={progress} max="100" />
+					{#if $telemetry.download_ota.status == 'progress'}
+						<progress class="progress progress-primary w-56" value={progress} max="100" />
+					{:else}
+						<progress class="progress progress-primary w-56" />
+					{/if}
 					<p class="mt-8 text-2xl">{message}</p>
 				</div>
 			</div>